backend/requirements.txt
numpy~=1.21.4
pandas~=1.3.4
requests~=2.26.0
beautifulsoup4~=4.10.0
SQLAlchemy~=1.4.27
MarkupSafe~=2.0.1
Flask~=2.0.2
psycopg2-binary~=2.9.2